Upsolved

+++ CodeForces/1015/1015E2.cpp
@@ -1,7 +1,7 @@
 // https://github.com/AkshatAggarwal14/Competetive-Programming
 #include "bits/stdc++.h"
 using namespace std;
-using ll = int64_t;
+using ll = int;
 auto sz = [](const auto &container) -> ll { return container.size(); };
 
 int main() {
